Move modjy's PyFilterTest and PyServletTest to their new package, too
