+++ tests/modjy/java/org/python/sevlet/PyFilterTest.java
@@ -1,4 +1,4 @@
-package org.python.util;
+package org.python.sevlet;
 
 import jakarta.servlet.http.HttpServletResponse;
